# vv/testdata/api/api/dummy.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "dummy.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "DummyService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/dummy/echo": {
      "get": {
        "operationId": "DummyService_Echo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dummyEchoResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DummyService"
        ]
      }
    },
    "/dummy/excel": {
      "get": {
        "operationId": "DummyService_Excel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dummyExcelResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "DummyService"
        ]
      }
    },
    "/dummy/picture": {
      "get": {
        "operationId": "DummyService_Picture",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dummyPictureResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "file_name",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DummyService"
        ]
      }
    },
    "/dummy/stream/echo": {
      "get": {
        "operationId": "DummyService_StreamEcho",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/dummyEchoResp"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of dummyEchoResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DummyService"
        ]
      }
    },
    "/dummy/upload/{file_name}": {
      "post": {
        "operationId": "DummyService_Upload",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dummyUploadResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "file_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string",
              "format": "byte"
            }
          }
        ],
        "tags": [
          "DummyService"
        ]
      }
    }
  },
  "definitions": {
    "dummyEchoResp": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "ack": {
          "type": "boolean"
        }
      }
    },
    "dummyExcelResp": {
      "type": "object",
      "properties": {
        "raw": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "dummyPictureResp": {
      "type": "object",
      "properties": {
        "raw": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "dummyUploadResp": {
      "type": "object",
      "properties": {
        "digest": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}